requirements.txt
psycopg2-binary
aiogram
asyncpg
groq